fixed file upload error

+++ backend/middleware/imageUpload.js
@@ -25,9 +25,11 @@ const uploadFile = async (file, folderName) => {
         throw new Error(`There was an error uploading the file: ${file}`);
       }
 
-      //delete file from temp folder (ptm-uploads)
-      fs.rmSync(fileToUpload);
-
+      //delete file from temp folder (ptm-uploads) :: DEV MODE
+      if (process.env.NODE_ENV === "development") {
+        fs.rmSync(fileToUpload);
+      }
+      
       //return the url of the uploaded file
       return response.secure_url;
     } catch (error) {
